Module working docs updated
